implement stored procedure for Booking table and add constraints for cascading deletion
