+++ logs.log
@@ -1,2 +1,2 @@
-2025-03-25 11:08:16,047 Succesfully set up the connection
-2025-03-25 11:08:16,074 Table 6 on 2022-01-24 is already booked
+2025-03-26 09:55:32,630 Succesfully set up the connection
+2025-03-26 09:55:32,671 Cancel Booking
