+++ main.py
@@ -2,12 +2,10 @@ import logging
 import os
 
 import mysql.connector as connector
-from dotenv import load_dotenv
 from connection import set_up_connection
 from insert_data import insert_all_data
+from booking import add_booking_stored_procedure,update_booking_stored_procedure,cancel_booking_stored_procedure
 
-
-load_dotenv()
 logging.basicConfig(filename="logs.log", level=logging.INFO, format="%(asctime)s %(message)s", filemode="w")
 
 
@@ -222,79 +220,7 @@ def order_detail_parametrized_query():
     finally:
         cnx.close()
       
-def check_booking_stored_procedure():
-    cnx=set_up_connection()
-    try:
-        cursor=cnx.cursor()
-        sql_use_db="""
-            USE littlelemondb
-        """
-        cursor.execute(sql_use_db)
-        sql_drop_procedure="""
-            DROP PROCEDURE IF EXISTS CheckBooking;
-        """
-        cursor.execute(sql_drop_procedure)
-        sql_check_booking_stored_procedure="""
-            CREATE PROCEDURE CheckBooking(IN tableNO INT)
-            BEGIN
-                SELECT 
-                    CASE 
-                        WHEN tableNO in (Select TableNo FROM Booking) THEN CONCAT('Table ',tableNO,' is already booked')
-                        ELSE CONCAT('Table ',tableNO,' is not already booked')
-                    END as BookingStatus;
-            END
-        """
-        cursor.execute(sql_check_booking_stored_procedure)
-        cursor.callproc('CheckBooking',(3,))
-        results=next(cursor.stored_results())
-        dataset=results.fetchall()[0][0]
-        print(dataset)
-        logging.info(f"{dataset}")
-    except connector.Error as e:
-        logging.error(f"Problem with creating CheckBooking stored procedure: {e}")
-    finally:
-        cnx.close()
+
+
         
-        
-def add_valid_bboking_stored_procedure():
-    cnx=set_up_connection()
-    try:
-        cursor=cnx.cursor()
-        cursor.execute("USE littlelemondb")
-        sql_drop_procedure = """
-            DROP PROCEDURE IF EXISTS CheckValidBooking;
-        """
-        cursor.execute(sql_drop_procedure)
-        sql_add_valid_booking = """
-            CREATE PROCEDURE CheckValidBooking(IN bookingDate DATE, IN tableNO INT)
-            BEGIN
-                DECLARE validBooking BOOLEAN;
-                START TRANSACTION;
-                SELECT 
-                    COUNT(*) INTO validBooking
-                FROM Booking
-                WHERE BookingDate = bookingDate AND TableNo = tableNO;
-                
-                IF validBooking = 0 THEN
-                    INSERT INTO Booking (BookingID, BookingDate, TableNo)
-                    VALUES ((SELECT IFNULL(MAX(BookingID), 0) + 1 FROM Booking), bookingDate, tableNO);
-                    SELECT CONCAT('Table ',tableNO,' on',bookingDate,' is not already booked')
-                    COMMIT;
-                ELSE
-                    SELECT CONCAT('Table ',tableNO,' on ',bookingDate,' is already booked')
-                    ROLLBACK;
-                END IF;
-            END;
-        """
-        cursor.execute(sql_add_valid_booking)
-        cursor.callproc('CheckValidBooking',('2022-01-24',6))
-        results=next(cursor.stored_results())
-        dataset=results.fetchall()[0][0]
-        print(dataset)
-        logging.info(f"{dataset}")
-    except connector.Error as e:
-        logging.error(f"Problem with creating AddValidBooking stored procedure: {e}")
-    finally:
-        cnx.close()
-        
-        
+cancel_booking_stored_procedure()
